bevy_client: port web-parity voxel/jump/prefab features
Brings the Bevy client up to the web client's currently shipped local
feature surface:

- Space jump produces movement frame flag 0x04 (input + observe).
- Offline-local voxel truth: macro place/break, refined 8x8x8 microgrid,
  prefab placement, boundary snap, snapshot import/export, hotbar.
- Built-in prefabs: builtin_sphere / builtin_cylinder / builtin_stairs.
- Perspective 3D camera with orbit follow + Ctrl+wheel zoom.
- Screen-center voxel raycast against macro and micro AABBs, hit-face
  highlight, adjacent macro/micro derivation.
- Material/prefab hotbar place + break through center selection.
- Prefab boundary-snap preview via micro-wire gizmos with browser
  fallback reasons no_target_boundary / no_contact / empty_prefab.
- Stdio/headless CLI parity for voxel commands: place/break/hotbar/
  prefab_place/prefab_place_snap/world_export/world_import/world_save/
  world_load/jump.
- Voxel parity + CLI parity integration tests.

Verified locally: cargo fmt -- --check, cargo clippy -D warnings,
cargo test (96 unit + 7 integration).

+++ clients/bevy_client/src/world/local_player.rs
@@ -205,7 +205,7 @@ impl LocalPredictionRuntime {
             let idle_frame = MoveInputFrame {
                 seq: 0,
                 client_tick: current.tick + 1,
-                dt_ms: self.profile.fixed_dt_ms as u16,
+                dt_ms: self.profile.fixed_dt_ms,
                 input_dir: Vec2::ZERO,
                 speed_scale: 1.0,
                 movement_flags: crate::input::commands::MOVEMENT_FLAG_BRAKE,
